requirements.txt
algopy==0.5.7
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
arviz==0.11.4
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
cftime==1.6.0
charset-normalizer==2.0.7
click==7.1.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecat==2.1.1
dill==0.3.4
entrypoints==0.3
fastprogress==1.0.2
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.6.0
google-auth-oauthlib==0.5.0
graphviz==0.19.1
gsheetsdb==0.1.13.1
gspread==5.2.0
gunicorn==20.1.0
gym==0.7.4
httplib2==0.20.4
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.8
networkx==2.7.1
notebook==6.4.5
numdifftools==0.9.40
numexpr==2.8.1
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==22.0.4
plotly==5.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.1
pyglet==1.5.21
Pygments==2.10.0
# pymc3==3.11.5
pymdptoolbox==4.0b3
Pympler==1.0.1
pyparsing==3.0.4
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
smmap==5.0.0
statsmodels==0.13.0
streamlit==1.4.0
streamlit-ace==0.1.1
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
Theano-PyMC==1.1.2
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.0
xarray==2022.3.0
xlrd==2.0.1
